pyproject.toml
[tool.poetry]
name = "scan-chromebooks"
version = "0.1.0"
description = ""
authors = ["Wolfhound905 <aiden8green@gmail.com>"]
readme = "README.md"
packages = [{include = "scan_chromebooks"}]

[tool.poetry.dependencies]
python = "^3.10"
python-dotenv = "^0.21.0"
aiohttp = "^3.8.3"
termcolor = "^2.1.1"

[tool.poetry.group.dev.dependencies]
black = "^22.10.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
